fix: prevent duplicate retries on 2xx responses and unrecoverable errors
Ported from ebms-core@e958a850 to the restructured codebase.

- Convert DeliveryTaskManager from an interface to an abstract base class
  holding shared DAO/CPAManager/retry fields and a shouldRetryUnreliable()
  helper (moved off the DeliveryTaskManager location in the old tree).
- DAODeliveryTaskManager now extends it and delegates retry decisions to
  shouldRetryUnreliable(), which skips SUCCEEDED responses and
  EbMSUnrecoverableResponseException so they are not requeued.
- Add a 5-arg updateTask(...) carrying the Exception; DeliveryTaskHandler
  passes the caught exception through so retry decisions can inspect it.
- Widen autoRetryInterval to long (DeliveryTaskManagerConfig).
- Add DeliveryTaskManagerTest covering the retry/unretry decision matrix.

+++ core/src/main/java/nl/clockwork/ebms/client/delivery/handler/DeliveryTaskManagerConfig.java
@@ -38,7 +38,7 @@ public class DeliveryTaskManagerConfig
 	@Value("${deliveryTaskManager.nrAutoRetries}")
 	int nrAutoRetries;
 	@Value("${deliveryTaskManager.autoRetryInterval}")
-	int autoRetryInterval;
+	long autoRetryInterval;
 
 	@Bean
 	public DeliveryTaskManager deliveryTaskManager(DeliveryTaskDAO deliveryTaskDAO, EbMSDAO ebMSDAO, CPAManager cpaManager)
